streamingccl: rename streamingest to physical
Release note: none.
Epic: none.
