+++ pkg/BUILD.bazel
@@ -98,7 +98,7 @@ ALL_TESTS = [
     "//pkg/ccl/streamingccl/logical:logical_test",
     "//pkg/ccl/streamingccl/replicationutils:replicationutils_test",
     "//pkg/ccl/streamingccl/streamclient:streamclient_test",
-    "//pkg/ccl/streamingccl/streamingest:streamingest_test",
+    "//pkg/ccl/streamingccl/physical:physical_test",
     "//pkg/ccl/streamingccl/streamproducer:streamproducer_test",
     "//pkg/ccl/telemetryccl:telemetryccl_test",
     "//pkg/ccl/testccl/authccl:authccl_test",
@@ -967,8 +967,8 @@ GO_TARGETS = [
     "//pkg/ccl/streamingccl/replicationutils:replicationutils_test",
     "//pkg/ccl/streamingccl/streamclient:streamclient",
     "//pkg/ccl/streamingccl/streamclient:streamclient_test",
-    "//pkg/ccl/streamingccl/streamingest:streamingest",
-    "//pkg/ccl/streamingccl/streamingest:streamingest_test",
+    "//pkg/ccl/streamingccl/physical:physical",
+    "//pkg/ccl/streamingccl/physical:physical_test",
     "//pkg/ccl/streamingccl/streamproducer:streamproducer",
     "//pkg/ccl/streamingccl/streamproducer:streamproducer_test",
     "//pkg/ccl/streamingccl:streamingccl",
