+++ pkg/ccl/BUILD.bazel
@@ -26,7 +26,7 @@ go_library(
         "//pkg/ccl/storageccl",
         "//pkg/ccl/storageccl/engineccl",
         "//pkg/ccl/streamingccl/logical",
-        "//pkg/ccl/streamingccl/streamingest",
+        "//pkg/ccl/streamingccl/physical",
         "//pkg/ccl/streamingccl/streamproducer",
         "//pkg/ccl/utilccl",
         "//pkg/ccl/workloadccl",
