+++ pkg/ccl/streamingccl/logical/BUILD.bazel
@@ -16,8 +16,8 @@ go_library(
         "//pkg/ccl/changefeedccl/cdcevent",
         "//pkg/ccl/changefeedccl/changefeedbase",
         "//pkg/ccl/streamingccl",
+        "//pkg/ccl/streamingccl/physical",
         "//pkg/ccl/streamingccl/streamclient",
-        "//pkg/ccl/streamingccl/streamingest",
         "//pkg/jobs",
         "//pkg/jobs/jobspb",
         "//pkg/jobs/jobsprofiler",
