+++ pkg/ccl/streamingccl/physical/alter_replication_job.go
@@ -6,7 +6,7 @@
 //
 //     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
 
-package streamingest
+package physical
 
 import (
 	"context"
