+++ pkg/ccl/streamingccl/physical/main_test.go
@@ -6,7 +6,7 @@
 //
 //     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
 
-package streamingest_test
+package physical_test
 
 import (
 	"os"
