+++ pkg/ccl/streamingccl/physical/merged_subscription_test.go
@@ -6,7 +6,7 @@
 //
 //     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
 
-package streamingest
+package physical
 
 import (
 	"context"
