+++ pkg/ccl/streamingccl/physical/testdata/alter_tenant
