+++ pkg/ccl/streamingccl/physical/testdata/initial_scan_span_configs
