+++ pkg/ccl/streamingccl/physical/testdata/show_without_pts
