+++ pkg/ccl/streamingccl/physical/testdata/span_configs
