+++ pkg/testutils/lint/lint_test.go
@@ -2468,9 +2468,9 @@ func TestLint(t *testing.T) {
 			":!ccl/serverccl/admin_test.go",
 			":!ccl/streamingccl/replicationtestutils/testutils.go",
 			":!ccl/streamingccl/streamclient/partitioned_stream_client_test.go",
-			":!ccl/streamingccl/streamingest/replication_random_client_test.go",
-			":!ccl/streamingccl/streamingest/stream_ingestion_job_test.go",
-			":!ccl/streamingccl/streamingest/stream_ingestion_processor_test.go",
+			":!ccl/streamingccl/physical/replication_random_client_test.go",
+			":!ccl/streamingccl/physical/stream_ingestion_job_test.go",
+			":!ccl/streamingccl/physical/stream_ingestion_processor_test.go",
 			":!ccl/streamingccl/streamproducer/producer_job_test.go",
 			":!ccl/streamingccl/streamproducer/replication_stream_test.go",
 			":!ccl/workloadccl/allccl/all_test.go",
